requirements.txt
albumentations==1.1.0
imageio==2.13.5
joblib==1.1.0
networkx==2.6.3
numpy==1.22.0
opencv-python-headless==4.5.5.62
packaging==21.3
pandas==1.3.5
Pillow==9.0.0
pyparsing==3.0.6
python-dateutil==2.8.2
pytz==2021.3
PyWavelets==1.2.0
PyYAML==6.0
qudida==0.0.4
scikit-image==0.19.1
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
sklearn==0.0
threadpoolctl==3.0.0
tifffile==2021.11.2
torch==1.10.1
torchvision==0.11.2
typing_extensions==4.0.1
